Lots of tests cleanup: - Rename AsyncKitTests to AsyncKitTestsCommon and remove do-nothing test case from it. Move the TestError enum there. - In all test cases simplify the repeated code (can't just de-duplicate 'cause it requires class inheritance which doesn't work well on Linux) - Rename FlattenTests.swift -> Collection+FlattenTests.swift - Add EventLoop+ConcurrencyTests, move performWithTask test to it, include test case for both versions, correctly guard usage. - Rename EventLoopGroup+Future.swift -> EventLoopGroup+FutureTests.swift - Rename FutureExtensionsTests.swift -> Future+MiscellaneousTests.swift and move the nonempty tests to Future+NonemptyTests.swift - Rename TransformTests.swift -> Future+TransformTests.swift
