--- Tests/AsyncKitTests/AsyncKitTests.swift
@@ -1,52 +1,0 @@
-import XCTest
-import AsyncKit
-import Logging
-
-final class AsyncKitTests: XCTestCase {
-    func testUniverseSanity() {
-        print(self.eventLoop)
-        XCTAssert(true)
-    }
-    
-    /// This TestCases EventLoopGroup
-    var group: EventLoopGroup!
-    
-    /// Returns the next EventLoop from the `group`
-    var eventLoop: EventLoop {
-        return self.group.next()
-    }
-    
-    /// Ensures logging is set up
-    override class func setUp() {
-        super.setUp()
-        XCTAssertTrue(isLoggingConfigured)
-    }
-    
-    /// Sets up the TestCase for use
-    /// and initializes the EventLoopGroup
-    override func setUpWithError() throws {
-        try super.setUpWithError()
-        self.group = MultiThreadedEventLoopGroup(numberOfThreads: 1)
-    }
-    
-    /// Tears down the TestCase and
-    /// shuts down the EventLoopGroup
-    override func tearDownWithError() throws {
-        try self.group.syncShutdownGracefully()
-        self.group = nil
-        try super.tearDownWithError()
-    }
-}
-
-func env(_ name: String) -> String? {
-    return ProcessInfo.processInfo.environment[name]
-}
-
-let isLoggingConfigured: Bool = {
-    LoggingSystem.bootstrap { label in
-        var handler = StreamLogHandler.standardOutput(label: label)
-        handler.logLevel = env("LOG_LEVEL").flatMap { Logger.Level(rawValue: $0) } ?? .debug
-        return handler
-    }
-    return true
-}()
