+++ Tests/AsyncKitTests/Future+TransformTests.swift
@@ -1,7 +1,7 @@
 import XCTest
 import AsyncKit
 
-class TransformTests: XCTestCase {
+class FutureTransformTests: XCTestCase {
     func testTransforms() throws {
         let future = eventLoop.makeSucceededFuture(Int.random(in: 0...100))
         
@@ -19,23 +19,14 @@ class TransformTests: XCTestCase {
         XCTAssert(try futureA.and(futureB).transform(to: futureBool).wait())
     }
 
-    /// This TestCases EventLoopGroup
     var group: EventLoopGroup!
-    
-    /// Returns the next EventLoop from the `group`
-    var eventLoop: EventLoop {
-        return self.group.next()
-    }
-    
-    /// Sets up the TestCase for use
-    /// and initializes the EventLoopGroup
+    var eventLoop: EventLoop { self.group.next() }
+
     override func setUpWithError() throws {
         try super.setUpWithError()
         self.group = MultiThreadedEventLoopGroup(numberOfThreads: 1)
     }
 
-    /// Tears down the TestCase and
-    /// shuts down the EventLoopGroup
     override func tearDownWithError() throws {
         try self.group.syncShutdownGracefully()
         self.group = nil
