+++ Tests/AsyncKitTests/Future+TryTests.swift
@@ -5,18 +5,14 @@ import XCTest
 final class FutureTryTests: XCTestCase {
     private let eventLoop = EmbeddedEventLoop()
 
-    private enum NIOError: Error {
-        case testError
-    }
-
     func testTryFlatMapPropagatesCallbackError() {
         let future = eventLoop.future(0)
             .tryFlatMap { _ -> EventLoopFuture<String> in
-                throw NIOError.testError
+                throw TestError.generic
             }
 
         XCTAssertThrowsError(try future.wait()) { error in
-            guard case NIOError.testError = error else {
+            guard case TestError.generic = error else {
                 XCTFail("Received an unexpected error: \(error)")
                 return
             }
@@ -26,11 +22,11 @@ final class FutureTryTests: XCTestCase {
     func testTryFlatMapPropagatesInnerError() {
         let future = eventLoop.future(0)
             .tryFlatMap { _ -> EventLoopFuture<String> in
-                self.eventLoop.makeFailedFuture(NIOError.testError)
+                self.eventLoop.makeFailedFuture(TestError.generic)
             }
 
         XCTAssertThrowsError(try future.wait()) { error in
-            guard case NIOError.testError = error else {
+            guard case TestError.generic = error else {
                 XCTFail("Received an unexpected error: \(error)")
                 return
             }
